mevbot/core/flash_loan.py
"""Flash loan integration for MEV strategies"""
from typing import Dict, List, Optional, Union, Tuple
from web3 import Web3
from eth_typing import Address
import asyncio
import logging
from dataclasses import dataclass
import numpy as np

logger = logging.getLogger(__name__)

# Supported flash loan providers
AAVE_V3 = "aave_v3"
DYDX = "dydx"
BALANCER = "balancer"
UNISWAP_V3 = "uniswap_v3"

@dataclass
class FlashLoanParams:
    provider: str
    token_address: Address
    amount: int
    callback_data: bytes
    expected_profit: float
    max_slippage: float = 0.005  # 0.5% default slippage

class FlashLoanManager:
    def __init__(self, w3: Web3, config: Dict):
        self.w3 = w3
        self.config = config
        
        # Initialize provider contracts
        self.providers = {
            AAVE_V3: self._init_aave(),
            DYDX: self._init_dydx(),
            BALANCER: self._init_balancer(),
            UNISWAP_V3: self._init_uniswap()
        }
        
        # Performance tracking
        self.total_loans = 0
        self.successful_loans = 0
        self.total_fees_paid = 0.0
        
    def _init_aave(self):
        """Initialize Aave V3 flash loan contract"""
        try:
            return self.w3.eth.contract(
                address=self.config['aave_lending_pool'],
                abi=self.config['aave_abi']
            )
        except Exception as e:
            logger.error(f"Error initializing Aave: {e}")
            return None
            
    def _init_dydx(self):
        """Initialize dYdX flash loan contract"""
        try:
            return self.w3.eth.contract(
                address=self.config['dydx_solo'],
                abi=self.config['dydx_abi']
            )
        except Exception as e:
            logger.error(f"Error initializing dYdX: {e}")
            return None
            
    def _init_balancer(self):
        """Initialize Balancer flash loan contract"""
        try:
            return self.w3.eth.contract(
                address=self.config['balancer_vault'],
                abi=self.config['balancer_abi']
            )
        except Exception as e:
            logger.error(f"Error initializing Balancer: {e}")
            return None
            
    def _init_uniswap(self):
        """Initialize Uniswap V3 flash loan contract"""
        try:
            return self.w3.eth.contract(
                address=self.config['uniswap_factory'],
                abi=self.config['uniswap_abi']
            )
        except Exception as e:
            logger.error(f"Error initializing Uniswap: {e}")
            return None
            
    async def get_optimal_provider(
        self,
        token: Address,
        amount: int
    ) -> Optional[str]:
        """Find the best flash loan provider based on fees and liquidity"""
        try:
            best_provider = None
            lowest_fee = float('inf')
            
            # Check each provider
            for provider in self.providers:
                if not self.providers[provider]:
                    continue
                    
                # Get provider fee and liquidity
                fee = await self._get_provider_fee(provider, token, amount)
                liquidity = await self._get_provider_liquidity(provider, token)
                
                if fee is None or liquidity is None:
                    continue
                    
                # Check if provider has sufficient liquidity
                if liquidity < amount:
                    continue
                    
                # Update best provider if fee is lower
                if fee < lowest_fee:
                    lowest_fee = fee
                    best_provider = provider
                    
            return best_provider
            
        except Exception as e:
            logger.error(f"Error finding optimal provider: {e}")
            return None
            
    async def _get_provider_fee(
        self,
        provider: str,
        token: Address,
        amount: int
    ) -> Optional[float]:
        """Get flash loan fee for a provider"""
        try:
            if provider == AAVE_V3:
                return amount * 0.0009  # 0.09% fee
            elif provider == DYDX:
                return 0  # No fee
            elif provider == BALANCER:
                return amount * 0.0001  # 0.01% fee
            elif provider == UNISWAP_V3:
                return amount * 0.0005  # 0.05% fee
            return None
            
        except Exception as e:
            logger.error(f"Error getting provider fee: {e}")
            return None
            
    async def _get_provider_liquidity(
        self,
        provider: str,
        token: Address
    ) -> Optional[int]:
        """Get available liquidity from a provider"""
        try:
            contract = self.providers[provider]
            if not contract:
                return None
                
            if provider == AAVE_V3:
                return await contract.functions.getReserveData(token).call()[0]
            elif provider == DYDX:
                return await contract.functions.getAccountWei(token).call()
            elif provider == BALANCER:
                return await contract.functions.getPoolTokenInfo(token).call()[1]
            elif provider == UNISWAP_V3:
                return await contract.functions.balanceOf(token).call()
                
            return None
            
        except Exception as e:
            logger.error(f"Error getting provider liquidity: {e}")
            return None
            
    async def prepare_loan(
        self,
        amount: int,
        token: str,
        route: str = None
    ) -> Optional[FlashLoanParams]:
        """Prepare flash loan parameters"""
        try:
            # Get optimal provider if route not specified
            provider = route or await self.get_optimal_provider(token, amount)
            if not provider:
                return None
                
            # Calculate expected profit (implementation specific)
            expected_profit = 0  # TODO: Implement profit calculation
            
            # Build flash loan parameters
            return FlashLoanParams(
                provider=provider,
                token_address=token,
                amount=amount,
                callback_data=b'',  # TODO: Implement callback data
                expected_profit=expected_profit
            )
            
        except Exception as e:
            logger.error(f"Error preparing flash loan: {str(e)}")
            return None
            
    async def execute_flash_loan(
        self,
        params: FlashLoanParams
    ) -> Tuple[bool, Optional[Dict]]:
        """Execute a flash loan transaction"""
        try:
            contract = self.providers[params.provider]
            if not contract:
                return False, None
                
            # Build flash loan transaction
            tx = await self._build_flash_loan_tx(params)
            if not tx:
                return False, None
                
            # Send transaction
            tx_hash = await self.w3.eth.send_transaction(tx)
            receipt = await self.w3.eth.wait_for_transaction_receipt(tx_hash)
            
            success = receipt['status'] == 1
            if success:
                self.successful_loans += 1
                self.total_fees_paid += await self._get_provider_fee(
                    params.provider,
                    params.token_address,
                    params.amount
                )
                
            self.total_loans += 1
            return success, receipt
            
        except Exception as e:
            logger.error(f"Error executing flash loan: {e}")
            return False, None
            
    async def _build_flash_loan_tx(
        self,
        params: FlashLoanParams
    ) -> Optional[Dict]:
        """Build flash loan transaction based on provider"""
        try:
            contract = self.providers[params.provider]
            if not contract:
                return None
                
            # Get current nonce
            nonce = await self.w3.eth.get_transaction_count(
                self.config['address']
            )
            
            # Build base transaction
            tx = {
                'from': self.config['address'],
                'nonce': nonce,
                'gas': 500000,  # Estimate
                'maxFeePerGas': await self._get_max_fee(),
                'maxPriorityFeePerGas': await self._get_priority_fee()
            }
            
            # Add provider-specific parameters
            if params.provider == AAVE_V3:
                tx.update({
                    'to': contract.address,
                    'data': contract.encodeABI(
                        fn_name='flashLoan',
                        args=[
                            self.config['address'],
                            [params.token_address],
                            [params.amount],
                            [0],  # Interest rate mode
                            self.config['address'],
                            params.callback_data,
                            0  # referralCode
                        ]
                    )
                })
            elif params.provider == DYDX:
                # Implement dYdX specific parameters
                pass
            elif params.provider == BALANCER:
                # Implement Balancer specific parameters
                pass
            elif params.provider == UNISWAP_V3:
                # Implement Uniswap specific parameters
                pass
                
            return tx
            
        except Exception as e:
            logger.error(f"Error building flash loan transaction: {e}")
            return None
            
    async def _get_max_fee(self) -> int:
        """Get current max fee per gas"""
        try:
            base_fee = (await self.w3.eth.get_block('latest'))['baseFeePerGas']
            return base_fee * 2
            
        except Exception as e:
            logger.error(f"Error getting max fee: {e}")
            return 50_000_000_000  # 50 GWEI default
            
    async def _get_priority_fee(self) -> int:
        """Get current priority fee"""
        try:
            return await self.w3.eth.max_priority_fee
            
        except Exception as e:
            logger.error(f"Error getting priority fee: {e}")
            return 2_000_000_000  # 2 GWEI default
